refactor tests to iterate on feature tests faster
