+++ src/test/cli.test.ts
@@ -5,71 +5,17 @@
 
 import * as assert from 'assert';
 import * as path from 'path';
-import * as cp from 'child_process';
 import * as fs from 'fs';
 import * as os from 'os';
+import { buildKitOptions, commandMarkerTests, devContainerDown, devContainerStop, devContainerUp, shellExec, UpResult } from './testUtils';
 
 const pkg = require('../../package.json');
-
-const buildKitOptions = [
-	{ text: 'non-BuildKit', options: { useBuildKit: false }, },
-	{ text: 'BuildKit', options: { useBuildKit: true }, },
-] as const;
-
-interface UpResult {
-	outcome: string;
-	containerId: string;
-	composeProjectName: string | undefined;
-}
 
 describe('Dev Containers CLI', function () {
 	this.timeout('120s');
 
 	const tmp = path.relative(process.cwd(), path.join(__dirname, 'tmp'));
 	const cli = `npx --prefix ${tmp} devcontainer`;
-	async function devContainerUp(workspaceFolder: string, options?: { useBuildKit?: boolean; userDataFolder?: string }): Promise<UpResult> {
-		const buildkitOption = (options?.useBuildKit ?? false) ? '' : ' --buildkit=never';
-		const userDataFolderOption = (options?.userDataFolder ?? false) ? ` --user-data-folder=${options?.userDataFolder}` : '';
-		const res = await shellExec(`${cli} up --workspace-folder ${workspaceFolder}${buildkitOption}${userDataFolderOption}`);
-		const response = JSON.parse(res.stdout);
-		assert.equal(response.outcome, 'success');
-		const { outcome, containerId, composeProjectName } = response as UpResult;
-		assert.ok(containerId, 'Container id not found.');
-		return { outcome, containerId, composeProjectName };
-	}
-	async function devContainerDown(options: { containerId?: string | null; composeProjectName?: string | null }) {
-		if (options.containerId) {
-			await shellExec(`docker rm -f ${options.containerId}`);
-		}
-		if (options.composeProjectName) {
-			await shellExec(`docker compose --project-name ${options.composeProjectName} down`);
-		}
-	}
-	async function devContainerStop(options: { containerId?: string | null; composeProjectName?: string | null }) {
-		if (options.containerId) {
-			await shellExec(`docker stop ${options.containerId}`);
-		}
-		if (options.composeProjectName) {
-			await shellExec(`docker compose --project-name ${options.composeProjectName} stop`);
-		}
-	}
-	async function pathExists(workspaceFolder: string, location: string) {
-		try {
-			await shellExec(`${cli} exec --workspace-folder ${workspaceFolder} test -e ${location}`);
-			return true;
-		} catch (err) {
-			return false;
-		}
-	}
-	async function commandMarkerTests(workspaceFolder: string, expected: { postCreate: boolean; postStart: boolean; postAttach: boolean }, message: string) {
-		const actual = {
-			postCreate: await pathExists(workspaceFolder, '/tmp/postCreateCommand.testmarker'),
-			postStart: await pathExists(workspaceFolder, '/tmp/postStartCommand.testmarker'),
-			postAttach: await pathExists(workspaceFolder, '/tmp/postAttachCommand.testmarker'),
-		};
-		assert.deepStrictEqual(actual, expected, message);
-	}
-
 
 	before('Install', async () => {
 		await shellExec(`rm -rf ${tmp}/node_modules`);
@@ -118,38 +64,6 @@ describe('Dev Containers CLI', function () {
 				console.log(res.stdout);
 				const response = JSON.parse(res.stdout);
 				assert.equal(response.outcome, 'success');
-			});
-
-			// Testing failing with invalid devcontainers
-			it('should fail when a non-existent v1 feature is in the config', async () => {
-				const testFolder = `${__dirname}/configs/invalid-configs/invalid-v1-features`;
-				const buildKitOption = (options?.useBuildKit ?? false) ? '' : ' --buildkit=never';
-				let success = false;
-				try {
-					await shellExec(`${cli} build --workspace-folder ${testFolder} ${buildKitOption} --log-level trace`);
-					success = true;
-				} catch (error) {
-					assert.equal(error.error.code, 1, 'Should fail with exit code 1');
-					const res = JSON.parse(error.stdout);
-					assert.equal(res.outcome, 'error');
-					assert.ok(res.message.indexOf('Failed to fetch tarball') > -1, `Actual error msg:  ${res.message}`);
-				}
-				assert.equal(success, false, 'expect non-successful call');
-			});
-			it('should fail when a non-existent v2 feature is in the config', async () => {
-				const testFolder = `${__dirname}/configs/invalid-configs/invalid-v2-features`;
-				const buildKitOption = (options?.useBuildKit ?? false) ? '' : ' --buildkit=never';
-				let success = false;
-				try {
-					await shellExec(`${cli} build --workspace-folder ${testFolder} ${buildKitOption} --log-level trace`);
-					success = true;
-				} catch (error) {
-					assert.equal(error.error.code, 1, 'Should fail with exit code 1');
-					const res = JSON.parse(error.stdout);
-					assert.equal(res.outcome, 'error');
-					assert.ok(res.message.indexOf('Failed to process feature') > -1, `Actual error msg:  ${res.message}`);
-				}
-				assert.equal(success, false, 'expect non-successful call');
 			});
 		});
 
@@ -305,7 +219,7 @@ describe('Dev Containers CLI', function () {
 			const testFolder = `${__dirname}/configs/compose-image-without-features`;
 			before(async () => {
 				// build and start the container
-				upResult = await devContainerUp(testFolder);
+				upResult = await devContainerUp(cli, testFolder);
 			});
 			after(async () => await devContainerDown({ composeProjectName: upResult?.composeProjectName }));
 			it('should succeed', () => {
@@ -317,7 +231,7 @@ describe('Dev Containers CLI', function () {
 			const testFolder = `${__dirname}/configs/compose-Dockerfile-without-features`;
 			before(async () => {
 				// build and start the container
-				upResult = await devContainerUp(testFolder);
+				upResult = await devContainerUp(cli, testFolder);
 			});
 			after(async () => await devContainerDown({ composeProjectName: upResult?.composeProjectName }));
 			it('should succeed', () => {
@@ -339,13 +253,13 @@ describe('Dev Containers CLI', function () {
 					userDataFolder = fs.mkdtempSync(path.join(tmpDir, 'dc-cli-test-'));
 
 					// build and start the container
-					upResult1 = await devContainerUp(testFolder, { userDataFolder });
+					upResult1 = await devContainerUp(cli, testFolder, { userDataFolder });
 
 					// stop the container but don't delete it
 					await shellExec(`docker compose --project-name ${upResult1.composeProjectName} stop`);
 
 					// restart the container
-					upResult2 = await devContainerUp(testFolder, { userDataFolder });
+					upResult2 = await devContainerUp(cli, testFolder, { userDataFolder });
 
 				});
 				after(async () => await devContainerDown({ composeProjectName: upResult2?.composeProjectName }));
@@ -373,7 +287,7 @@ describe('Dev Containers CLI', function () {
 					const userDataFolder = fs.mkdtempSync(path.join(tmpDir, 'dc-cli-test-'));
 
 					// build and start the container
-					upResult1 = await devContainerUp(testFolder, { userDataFolder });
+					upResult1 = await devContainerUp(cli, testFolder, { userDataFolder });
 
 					// stop the container but don't delete it
 					await shellExec(`docker compose --project-name ${upResult1.composeProjectName} stop`);
@@ -384,7 +298,7 @@ describe('Dev Containers CLI', function () {
 					fs.mkdirSync(userDataFolder);
 
 					// restart the container
-					upResult2 = await devContainerUp(testFolder, { userDataFolder });
+					upResult2 = await devContainerUp(cli, testFolder, { userDataFolder });
 
 				});
 				after(async () => await devContainerDown({ composeProjectName: upResult2?.composeProjectName }));
@@ -402,7 +316,7 @@ describe('Dev Containers CLI', function () {
 		describe('with valid config', () => {
 			let containerId: string | null = null;
 			const testFolder = `${__dirname}/configs/image`;
-			beforeEach(async () => containerId = (await devContainerUp(testFolder)).containerId);
+			beforeEach(async () => containerId = (await devContainerUp(cli, testFolder)).containerId);
 			afterEach(async () => await devContainerDown({ containerId }));
 			it('should execute successfully', async () => {
 				const res = await shellExec(`${cli} run-user-commands --workspace-folder ${testFolder}`);
@@ -432,7 +346,7 @@ describe('Dev Containers CLI', function () {
 			describe(`with valid (image) config [${text}]`, () => {
 				let containerId: string | null = null;
 				const testFolder = `${__dirname}/configs/image`;
-				beforeEach(async () => containerId = (await devContainerUp(testFolder, options)).containerId);
+				beforeEach(async () => containerId = (await devContainerUp(cli, testFolder, options)).containerId);
 				afterEach(async () => await devContainerDown({ containerId }));
 				it('should execute successfully', async () => {
 					const res = await shellExec(`${cli} exec --workspace-folder ${testFolder} echo hi`);
@@ -443,7 +357,7 @@ describe('Dev Containers CLI', function () {
 			describe(`with valid (image) config containing features [${text}]`, () => {
 				let containerId: string | null = null;
 				const testFolder = `${__dirname}/configs/image-with-features`;
-				beforeEach(async () => containerId = (await devContainerUp(testFolder, options)).containerId);
+				beforeEach(async () => containerId = (await devContainerUp(cli, testFolder, options)).containerId);
 				afterEach(async () => await devContainerDown({ containerId }));
 				it('should have access to installed features (docker)', async () => {
 					const res = await shellExec(`${cli} exec --workspace-folder ${testFolder} docker --version`);
@@ -463,7 +377,7 @@ describe('Dev Containers CLI', function () {
 			describe(`with valid (Dockerfile) config containing features [${text}]`, () => {
 				let containerId: string | null = null;
 				const testFolder = `${__dirname}/configs/dockerfile-with-features`;
-				beforeEach(async () => containerId = (await devContainerUp(testFolder, options)).containerId);
+				beforeEach(async () => containerId = (await devContainerUp(cli, testFolder, options)).containerId);
 				afterEach(async () => await devContainerDown({ containerId }));
 				it('should have access to installed features (docker)', async () => {
 					// NOTE: Doing a docker ps will ensure that the --privileged flag was set by the feature
@@ -485,7 +399,7 @@ describe('Dev Containers CLI', function () {
 			describe(`with valid (Dockerfile) config with target [${text}]`, () => {
 				let containerId: string | null = null;
 				const testFolder = `${__dirname}/configs/dockerfile-with-target`;
-				beforeEach(async () => containerId = (await devContainerUp(testFolder, options)).containerId);
+				beforeEach(async () => containerId = (await devContainerUp(cli, testFolder, options)).containerId);
 				afterEach(async () => await devContainerDown({ containerId }));
 				it('should have build marker content', async () => {
 					const res = await shellExec(`${cli} exec --workspace-folder ${testFolder} cat /var/test-marker`);
@@ -496,25 +410,10 @@ describe('Dev Containers CLI', function () {
 				});
 			});
 
-			describe(`with valid (Dockerfile) config and v2 OCI feature (dind) [${text}]`, () => {
-				let containerId: string | null = null;
-				const testFolder = `${__dirname}/configs/dockerfile-with-v2-oci-features`;
-				beforeEach(async () => containerId = (await devContainerUp(testFolder, options)).containerId);
-				afterEach(async () => await devContainerDown({ containerId }));
-				it('should detect docker installed (--privileged flag passed)', async () => {
-					// NOTE: Doing a docker ps will ensure that the --privileged flag was set by the feature
-					const res = await shellExec(`${cli} exec --workspace-folder ${testFolder} docker ps`);
-					const response = JSON.parse(res.stdout);
-					console.log(res.stderr);
-					assert.equal(response.outcome, 'success');
-					assert.match(res.stderr, /CONTAINER ID/);
-				});
-			});
-
 			describe(`with valid (docker-compose with image) config containing v1 features [${text}]`, () => {
 				let composeProjectName: string | undefined = undefined;
 				const testFolder = `${__dirname}/configs/compose-image-with-features`;
-				beforeEach(async () => composeProjectName = (await devContainerUp(testFolder, options)).composeProjectName);
+				beforeEach(async () => composeProjectName = (await devContainerUp(cli, testFolder, options)).composeProjectName);
 				afterEach(async () => await devContainerDown({ composeProjectName }));
 				it('should have access to installed features (docker)', async () => {
 					// NOTE: Doing a docker ps will ensure that the --privileged flag was set by the feature
@@ -536,7 +435,7 @@ describe('Dev Containers CLI', function () {
 			describe(`with valid (docker-compose with Dockerfile) config containing features [${text}]`, () => {
 				let composeProjectName: string | undefined = undefined;
 				const testFolder = `${__dirname}/configs/compose-Dockerfile-with-features`;
-				beforeEach(async () => composeProjectName = (await devContainerUp(testFolder, options)).composeProjectName);
+				beforeEach(async () => composeProjectName = (await devContainerUp(cli, testFolder, options)).composeProjectName);
 				afterEach(async () => await devContainerDown({ composeProjectName }));
 				it('should have access to installed features (docker)', async () => {
 					const res = await shellExec(`${cli} exec --workspace-folder ${testFolder} docker --version`);
@@ -557,7 +456,7 @@ describe('Dev Containers CLI', function () {
 			describe(`with valid (docker-compose with Dockerfile and target) config containing features [${text}]`, () => {
 				let composeProjectName: string | undefined = undefined;
 				const testFolder = `${__dirname}/configs/compose-Dockerfile-with-target`;
-				beforeEach(async () => composeProjectName = (await devContainerUp(testFolder, options)).composeProjectName);
+				beforeEach(async () => composeProjectName = (await devContainerUp(cli, testFolder, options)).composeProjectName);
 				afterEach(async () => await devContainerDown({ composeProjectName }));
 				it('should have access to installed features (docker)', async () => {
 					const res = await shellExec(`${cli} exec --workspace-folder ${testFolder} docker --version`);
@@ -588,17 +487,17 @@ describe('Dev Containers CLI', function () {
 				const testFolder = `${__dirname}/configs/dockerfile-with-target`;
 				after(async () => await devContainerDown({ containerId }));
 				it('should have all command markers at appropriate times', async () => {
-					containerId = (await devContainerUp(testFolder, options)).containerId;
+					containerId = (await devContainerUp(cli, testFolder, options)).containerId;
 					// Should have all markers (Create + Start + Attach)
-					await commandMarkerTests(testFolder, { postCreate: true, postStart: true, postAttach: true }, 'Markers on first create');
+					await commandMarkerTests(cli, testFolder, { postCreate: true, postStart: true, postAttach: true }, 'Markers on first create');
 
 					// Clear markers and stop
 					await shellExec(`${cli} exec --workspace-folder ${testFolder} /bin/sh -c "rm /tmp/*.testmarker"`);
 					await devContainerStop({ containerId });
 
 					// Restart container - should have Start + Attach
-					containerId = (await devContainerUp(testFolder, options)).containerId;
-					await commandMarkerTests(testFolder, { postCreate: false, postStart: true, postAttach: true }, 'Markers on restart');
+					containerId = (await devContainerUp(cli, testFolder, options)).containerId;
+					await commandMarkerTests(cli, testFolder, { postCreate: false, postStart: true, postAttach: true }, 'Markers on restart');
 
 					// TODO - investigate what triggers postAttachCommand and check test is valid
 					// // Clear markers and re-test - should have Attach
@@ -611,17 +510,17 @@ describe('Dev Containers CLI', function () {
 				const testFolder = `${__dirname}/configs/compose-Dockerfile-with-target`;
 				after(async () => await devContainerDown({ composeProjectName }));
 				it('should have all command markers at appropriate times', async () => {
-					composeProjectName = (await devContainerUp(testFolder, options)).composeProjectName;
+					composeProjectName = (await devContainerUp(cli, testFolder, options)).composeProjectName;
 					// Should have all markers (Create + Start + Attach)
-					await commandMarkerTests(testFolder, { postCreate: true, postStart: true, postAttach: true }, 'Markers on first create');
+					await commandMarkerTests(cli, testFolder, { postCreate: true, postStart: true, postAttach: true }, 'Markers on first create');
 
 					// Clear markers and stop
 					await shellExec(`${cli} exec --workspace-folder ${testFolder} /bin/sh -c "rm /tmp/*.testmarker"`);
 					await devContainerStop({ composeProjectName });
 
 					// Restart container - should have Start + Attach
-					composeProjectName = (await devContainerUp(testFolder, options)).composeProjectName;
-					await commandMarkerTests(testFolder, { postCreate: false, postStart: true, postAttach: true }, 'Markers on restart');
+					composeProjectName = (await devContainerUp(cli, testFolder, options)).composeProjectName;
+					await commandMarkerTests(cli, testFolder, { postCreate: false, postStart: true, postAttach: true }, 'Markers on restart');
 
 					// TODO - investigate what triggers postAttachCommand and check test is valid
 					// // Clear markers and re-test - should have Attach
@@ -634,17 +533,17 @@ describe('Dev Containers CLI', function () {
 				const testFolder = `${__dirname}/configs/compose-Dockerfile-with-target`;
 				after(async () => await devContainerDown({ containerId }));
 				it('should have all command markers at appropriate times', async () => {
-					containerId = (await devContainerUp(testFolder, options)).containerId;
+					containerId = (await devContainerUp(cli, testFolder, options)).containerId;
 					// Should have all markers (Create + Start + Attach)
-					await commandMarkerTests(testFolder, { postCreate: true, postStart: true, postAttach: true }, 'Markers on first create');
+					await commandMarkerTests(cli, testFolder, { postCreate: true, postStart: true, postAttach: true }, 'Markers on first create');
 
 					// Clear markers and stop
 					await shellExec(`${cli} exec --workspace-folder ${testFolder} /bin/sh -c "rm /tmp/*.testmarker"`);
 					await devContainerStop({ containerId });
 
 					// Restart container - should have Start + Attach
-					containerId = (await devContainerUp(testFolder, options)).containerId;
-					await commandMarkerTests(testFolder, { postCreate: false, postStart: true, postAttach: true }, 'Markers on restart');
+					containerId = (await devContainerUp(cli, testFolder, options)).containerId;
+					await commandMarkerTests(cli, testFolder, { postCreate: false, postStart: true, postAttach: true }, 'Markers on restart');
 
 					// TODO - investigate what triggers postAttachCommand and check test is valid
 					// // Clear markers and re-test - should have Attach
@@ -656,9 +555,9 @@ describe('Dev Containers CLI', function () {
 					const testFolder = `${__dirname}/configs/compose-Dockerfile-alpine`;
 					after(async () => await devContainerDown({ containerId }));
 					it('should have all command markers at appropriate times', async () => {
-						containerId = (await devContainerUp(testFolder, options)).containerId;
+						containerId = (await devContainerUp(cli, testFolder, options)).containerId;
 						// Should have all markers (Create + Start + Attach)
-						await commandMarkerTests(testFolder, { postCreate: true, postStart: true, postAttach: true }, 'Markers on first create');
+						await commandMarkerTests(cli, testFolder, { postCreate: true, postStart: true, postAttach: true }, 'Markers on first create');
 
 						// Clear markers and stop
 						await shellExec(`${cli} exec --workspace-folder ${testFolder} /bin/sh -c "rm /tmp/*.testmarker"`);
@@ -666,8 +565,8 @@ describe('Dev Containers CLI', function () {
 						await devContainerStop({ containerId });
 
 						// Restart container - should have Start + Attach
-						containerId = (await devContainerUp(testFolder, options)).containerId;
-						await commandMarkerTests(testFolder, { postCreate: false, postStart: true, postAttach: true }, 'Markers on restart');
+						containerId = (await devContainerUp(cli, testFolder, options)).containerId;
+						await commandMarkerTests(cli, testFolder, { postCreate: false, postStart: true, postAttach: true }, 'Markers on restart');
 
 						// TODO - investigate what triggers postAttachCommand and check test is valid
 						// // Clear markers and re-test - should have Attach
@@ -681,7 +580,7 @@ describe('Dev Containers CLI', function () {
 		describe('with valid (Dockerfile) config containing #syntax (BuildKit)', () => { // ensure existing '# syntax' lines are handled
 			let containerId: string | null = null;
 			const testFolder = `${__dirname}/configs/dockerfile-with-syntax`;
-			beforeEach(async () => containerId = (await devContainerUp(testFolder, { useBuildKit: true })).containerId);
+			beforeEach(async () => containerId = (await devContainerUp(cli, testFolder, { useBuildKit: true })).containerId);
 			afterEach(async () => await devContainerDown({ containerId }));
 			it('should have access to installed features (docker)', async () => {
 				const res = await shellExec(`${cli} exec --workspace-folder ${testFolder} docker --version`);
@@ -712,27 +611,5 @@ describe('Dev Containers CLI', function () {
 			}
 			assert.equal(success, false, 'expect non-successful call');
 		});
-
-
 	});
-
-
 });
-
-interface ExecResult {
-	error: Error | null;
-	stdout: string;
-	stderr: string;
-}
-
-function shellExec(command: string, options: cp.ExecOptions = {}, suppressOutput: boolean = false) {
-	return new Promise<ExecResult>((resolve, reject) => {
-		cp.exec(command, options, (error, stdout, stderr) => {
-			if (!suppressOutput) {
-				console.log(stdout);
-				console.error(stderr);
-			}
-			(error ? reject : resolve)({ error, stdout, stderr });
-		});
-	});
-}
