+++ src/test/container-features/configs/dockerfile-with-v2-oci-features/.devcontainer.json
