+++ src/test/container-features/configs/invalid-configs/invalid-v1-features/.devcontainer.json
