+++ src/test/container-features/configs/invalid-configs/invalid-v2-features/.devcontainer.json
