+++ src/test/container-features/example-v1-features-sets/simple/install.sh
